+++ lib/Braintree/CreditCardGateway.php
@@ -309,7 +309,24 @@ class CreditCardGateway
         $options[] = "failOnDuplicatePaymentMethod";
         $signature = self::baseSignature($options);
         $signature[] = 'customerId';
+        $signature[] = self::threeDSecurePassThruSignature();
         return $signature;
+    }
+
+    public static function threeDSecurePassThruSignature()
+    {
+        return [
+            'threeDSecurePassThru' => [
+                'eciFlag',
+                'cavv',
+                'xid',
+                'threeDSecureVersion',
+                'authenticationResponse',
+                'directoryResponse',
+                'cavvAlgorithm',
+                'dsTransactionId',
+            ]
+        ];
     }
 
     public static function updateSignature()
@@ -317,6 +334,7 @@ class CreditCardGateway
         $options = self::baseOptions();
         $options[] = "failOnDuplicatePaymentMethod";
         $signature = self::baseSignature($options);
+        $signature[] = self::threeDSecurePassThruSignature();
 
         $updateExistingBillingSignature = [
             [