+++ lib/Braintree/CustomerGateway.php
@@ -111,7 +111,6 @@ class CustomerGateway
     {
         $creditCardSignature = CreditCardGateway::createSignature();
         unset($creditCardSignature[array_search('customerId', $creditCardSignature)]);
-
         $signature = [
             'id', 'company', 'email', 'fax', 'firstName',
             'lastName', 'phone', 'website', 'deviceData',
@@ -143,7 +142,7 @@ class CustomerGateway
         ];
         return $signature;
     }
-
+ 
     /**
      * creates a full array signature of a valid update request
      * @return array update request format