+++ CHANGELOG.md
@@ -8,7 +8,6 @@
   * Remove deprecated SEPA error codes
   * Remove `paypalVaultWithoutUpgrade`
   * Invalid transaction IDs are validated by the gateway and not the SDK
-  * Require PHP 7.2 or higher
 
 ## 4.8.0
 * Add `Authentication Insight` to payment method nonce create
@@ -70,6 +69,7 @@
 
 ## 4.0.0
 * Add support for PHP 7 (Thanks to @briandotdev)
+* Require PHP 7.2 or higher
 * Remove support for HHVM and PHP 5
 * Update to PHPUnit 7
 * Remove deprecated `GRANTED_PAYMENT_INSTRUMENT_UPDATE`