+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## unreleased
+* Add `threeDSecurePassThru` parameters to `Customer.create`, `PaymentMethod.create`, `CreditCard.create`, `Customer.update`, `PaymentMethod.update` and `CreditCard.update` 
 * Add `recurringCustomerConsent` and `recurringMaxAmount` to `authenticationInsightOptions` for `PaymentMethodNonce.create`
 * Add `DOCUMENT_UPLOAD_FILE_IS_EMPTY` error code